refactor: organize the ui code hierarchically
