+++ alt-tab-macos/ui/Menubar.swift
@@ -1,6 +1,6 @@
 import Cocoa
 
-class StatusItem {
+class Menubar {
     static func make(_ app: App) -> NSStatusItem {
         let item = NSStatusBar.system.statusItem(withLength: NSStatusItem.variableLength)
         item.button!.title = App.name
