+++ alt-tab-macos/ui/labels/FontIcon.swift
